flask_product_api/requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pytz==2023.3
six==1.16.0
Werkzeug==2.2.3
python-dotenv
